refactor: split role service methods
